+++ packages/orq-rc/src/orq_ai_sdk/knowledge.py
@@ -5,7 +5,7 @@ from orq_ai_sdk import models, utils
 from orq_ai_sdk._hooks import HookContext
 from orq_ai_sdk.types import OptionalNullable, UNSET
 from orq_ai_sdk.utils import get_security_from_env
-from typing import Any, List, Mapping, Optional, Union
+from typing import List, Mapping, Optional, Union
 
 
 class Knowledge(BaseSDK):
@@ -221,7 +221,7 @@ class Knowledge(BaseSDK):
         self,
         *,
         key: str,
-        model: str,
+        embedding_model: str,
         path: str,
         description: Optional[str] = None,
         retrieval_settings: Optional[
@@ -235,7 +235,7 @@ class Knowledge(BaseSDK):
         r"""Create a knowledge
 
         :param key:
-        :param model: The embeddings model to use for the knowledge base.
+        :param embedding_model: The embeddings model to use for the knowledge base. This model will be used to embed the chunks when they are added to the knowledge base.
         :param path: The path where the entity is stored in the project structure. The first element of the path always represents the project name. Any subsequent path element after the project will be created as a folder in the project if it does not exists.
         :param description:
         :param retrieval_settings: The retrieval settings for the knowledge base. If not provider, Hybrid Search will be used as a default query strategy.
@@ -260,7 +260,7 @@ class Knowledge(BaseSDK):
         request = models.CreateKnowledgeRequestBody(
             key=key,
             description=description,
-            model=model,
+            embedding_model=embedding_model,
             retrieval_settings=utils.get_pydantic_model(
                 retrieval_settings, Optional[models.RetrievalSettings]
             ),
@@ -336,7 +336,7 @@ class Knowledge(BaseSDK):
         self,
         *,
         key: str,
-        model: str,
+        embedding_model: str,
         path: str,
         description: Optional[str] = None,
         retrieval_settings: Optional[
@@ -350,7 +350,7 @@ class Knowledge(BaseSDK):
         r"""Create a knowledge
 
         :param key:
-        :param model: The embeddings model to use for the knowledge base.
+        :param embedding_model: The embeddings model to use for the knowledge base. This model will be used to embed the chunks when they are added to the knowledge base.
         :param path: The path where the entity is stored in the project structure. The first element of the path always represents the project name. Any subsequent path element after the project will be created as a folder in the project if it does not exists.
         :param description:
         :param retrieval_settings: The retrieval settings for the knowledge base. If not provider, Hybrid Search will be used as a default query strategy.
@@ -375,7 +375,7 @@ class Knowledge(BaseSDK):
         request = models.CreateKnowledgeRequestBody(
             key=key,
             description=description,
-            model=model,
+            embedding_model=embedding_model,
             retrieval_settings=utils.get_pydantic_model(
                 retrieval_settings, Optional[models.RetrievalSettings]
             ),
@@ -648,7 +648,7 @@ class Knowledge(BaseSDK):
         *,
         knowledge_id: str,
         description: Optional[str] = None,
-        model: Optional[str] = None,
+        embedding_model: Optional[str] = None,
         path: Optional[str] = None,
         retrieval_settings: Optional[
             Union[
@@ -665,9 +665,9 @@ class Knowledge(BaseSDK):
 
         :param knowledge_id: The unique identifier of the knowledge base
         :param description: The description of the knowledge base.
-        :param model: The embeddings model used for the knowledge base. If the models is provided and is different than the previous set model, all the datasources in the knowledge base will be re-embedded.
+        :param embedding_model: The embeddings model used for the knowledge base. If the models is provided and is different than the previous set model, all the datasources in the knowledge base will be re-embedded.
         :param path: The path where the entity is stored in the project structure. The first element of the path always represents the project name. Any subsequent path element after the project will be created as a folder in the project if it does not exists.
-        :param retrieval_settings:
+        :param retrieval_settings: The retrieval settings for the knowledge base. If not provider, Hybrid Search will be used as a default query strategy.
         :param retries: Override the default retry configuration for this method
         :param server_url: Override the default server URL for this method
         :param timeout_ms: Override the default request timeout configuration for this method in milliseconds
@@ -690,7 +690,7 @@ class Knowledge(BaseSDK):
             knowledge_id=knowledge_id,
             request_body=models.UpdateKnowledgeRequestBody(
                 description=description,
-                model=model,
+                embedding_model=embedding_model,
                 path=path,
                 retrieval_settings=utils.get_pydantic_model(
                     retrieval_settings,
@@ -773,7 +773,7 @@ class Knowledge(BaseSDK):
         *,
         knowledge_id: str,
         description: Optional[str] = None,
-        model: Optional[str] = None,
+        embedding_model: Optional[str] = None,
         path: Optional[str] = None,
         retrieval_settings: Optional[
             Union[
@@ -790,9 +790,9 @@ class Knowledge(BaseSDK):
 
         :param knowledge_id: The unique identifier of the knowledge base
         :param description: The description of the knowledge base.
-        :param model: The embeddings model used for the knowledge base. If the models is provided and is different than the previous set model, all the datasources in the knowledge base will be re-embedded.
+        :param embedding_model: The embeddings model used for the knowledge base. If the models is provided and is different than the previous set model, all the datasources in the knowledge base will be re-embedded.
         :param path: The path where the entity is stored in the project structure. The first element of the path always represents the project name. Any subsequent path element after the project will be created as a folder in the project if it does not exists.
-        :param retrieval_settings:
+        :param retrieval_settings: The retrieval settings for the knowledge base. If not provider, Hybrid Search will be used as a default query strategy.
         :param retries: Override the default retry configuration for this method
         :param server_url: Override the default server URL for this method
         :param timeout_ms: Override the default request timeout configuration for this method in milliseconds
@@ -815,7 +815,7 @@ class Knowledge(BaseSDK):
             knowledge_id=knowledge_id,
             request_body=models.UpdateKnowledgeRequestBody(
                 description=description,
-                model=model,
+                embedding_model=embedding_model,
                 path=path,
                 retrieval_settings=utils.get_pydantic_model(
                     retrieval_settings,
@@ -1296,224 +1296,6 @@ class Knowledge(BaseSDK):
                 http_res.text, Optional[models.SearchKnowledgeResponseBody]
             )
         if utils.match_response(http_res, ["404", "4XX"], "*"):
-            http_res_text = await utils.stream_to_text_async(http_res)
-            raise models.APIError(
-                "API error occurred", http_res.status_code, http_res_text, http_res
-            )
-        if utils.match_response(http_res, "5XX", "*"):
-            http_res_text = await utils.stream_to_text_async(http_res)
-            raise models.APIError(
-                "API error occurred", http_res.status_code, http_res_text, http_res
-            )
-
-        content_type = http_res.headers.get("Content-Type")
-        http_res_text = await utils.stream_to_text_async(http_res)
-        raise models.APIError(
-            f"Unexpected response received (code: {http_res.status_code}, type: {content_type})",
-            http_res.status_code,
-            http_res_text,
-            http_res,
-        )
-
-    def retrieve_file_url(
-        self,
-        *,
-        knowledge_id: str,
-        file_name: str,
-        content_type: str,
-        datasource_id: str,
-        retries: OptionalNullable[utils.RetryConfig] = UNSET,
-        server_url: Optional[str] = None,
-        timeout_ms: Optional[int] = None,
-        http_headers: Optional[Mapping[str, str]] = None,
-    ) -> Optional[models.GetOneFileUploadURLResponseBody]:
-        r"""Retrieves the file upload url
-
-        :param knowledge_id: The unique identifier of the knowledge base
-        :param file_name:
-        :param content_type:
-        :param datasource_id:
-        :param retries: Override the default retry configuration for this method
-        :param server_url: Override the default server URL for this method
-        :param timeout_ms: Override the default request timeout configuration for this method in milliseconds
-        :param http_headers: Additional headers to set or replace on requests.
-        """
-        base_url = None
-        url_variables = None
-        if timeout_ms is None:
-            timeout_ms = self.sdk_configuration.timeout_ms
-
-        if timeout_ms is None:
-            timeout_ms = 600000
-
-        if server_url is not None:
-            base_url = server_url
-        else:
-            base_url = self._get_url(base_url, url_variables)
-
-        request = models.GetOneFileUploadURLRequest(
-            knowledge_id=knowledge_id,
-            file_name=file_name,
-            content_type=content_type,
-            datasource_id=datasource_id,
-        )
-
-        req = self._build_request(
-            method="GET",
-            path="/v2/knowledge/{knowledge_id}/upload-file",
-            base_url=base_url,
-            url_variables=url_variables,
-            request=request,
-            request_body_required=False,
-            request_has_path_params=True,
-            request_has_query_params=True,
-            user_agent_header="user-agent",
-            accept_header_value="application/json",
-            http_headers=http_headers,
-            security=self.sdk_configuration.security,
-            timeout_ms=timeout_ms,
-        )
-
-        if retries == UNSET:
-            if self.sdk_configuration.retry_config is not UNSET:
-                retries = self.sdk_configuration.retry_config
-
-        retry_config = None
-        if isinstance(retries, utils.RetryConfig):
-            retry_config = (retries, ["429", "500", "502", "503", "504"])
-
-        http_res = self.do_request(
-            hook_ctx=HookContext(
-                base_url=base_url or "",
-                operation_id="GetOneFileUploadUrl",
-                oauth2_scopes=[],
-                security_source=get_security_from_env(
-                    self.sdk_configuration.security, models.Security
-                ),
-            ),
-            request=req,
-            error_status_codes=["4XX", "500", "5XX"],
-            retry_config=retry_config,
-        )
-
-        response_data: Any = None
-        if utils.match_response(http_res, "200", "application/json"):
-            return utils.unmarshal_json(
-                http_res.text, Optional[models.GetOneFileUploadURLResponseBody]
-            )
-        if utils.match_response(http_res, "500", "application/json"):
-            response_data = utils.unmarshal_json(http_res.text, models.HonoAPIErrorData)
-            raise models.HonoAPIError(data=response_data)
-        if utils.match_response(http_res, "4XX", "*"):
-            http_res_text = utils.stream_to_text(http_res)
-            raise models.APIError(
-                "API error occurred", http_res.status_code, http_res_text, http_res
-            )
-        if utils.match_response(http_res, "5XX", "*"):
-            http_res_text = utils.stream_to_text(http_res)
-            raise models.APIError(
-                "API error occurred", http_res.status_code, http_res_text, http_res
-            )
-
-        content_type = http_res.headers.get("Content-Type")
-        http_res_text = utils.stream_to_text(http_res)
-        raise models.APIError(
-            f"Unexpected response received (code: {http_res.status_code}, type: {content_type})",
-            http_res.status_code,
-            http_res_text,
-            http_res,
-        )
-
-    async def retrieve_file_url_async(
-        self,
-        *,
-        knowledge_id: str,
-        file_name: str,
-        content_type: str,
-        datasource_id: str,
-        retries: OptionalNullable[utils.RetryConfig] = UNSET,
-        server_url: Optional[str] = None,
-        timeout_ms: Optional[int] = None,
-        http_headers: Optional[Mapping[str, str]] = None,
-    ) -> Optional[models.GetOneFileUploadURLResponseBody]:
-        r"""Retrieves the file upload url
-
-        :param knowledge_id: The unique identifier of the knowledge base
-        :param file_name:
-        :param content_type:
-        :param datasource_id:
-        :param retries: Override the default retry configuration for this method
-        :param server_url: Override the default server URL for this method
-        :param timeout_ms: Override the default request timeout configuration for this method in milliseconds
-        :param http_headers: Additional headers to set or replace on requests.
-        """
-        base_url = None
-        url_variables = None
-        if timeout_ms is None:
-            timeout_ms = self.sdk_configuration.timeout_ms
-
-        if timeout_ms is None:
-            timeout_ms = 600000
-
-        if server_url is not None:
-            base_url = server_url
-        else:
-            base_url = self._get_url(base_url, url_variables)
-
-        request = models.GetOneFileUploadURLRequest(
-            knowledge_id=knowledge_id,
-            file_name=file_name,
-            content_type=content_type,
-            datasource_id=datasource_id,
-        )
-
-        req = self._build_request_async(
-            method="GET",
-            path="/v2/knowledge/{knowledge_id}/upload-file",
-            base_url=base_url,
-            url_variables=url_variables,
-            request=request,
-            request_body_required=False,
-            request_has_path_params=True,
-            request_has_query_params=True,
-            user_agent_header="user-agent",
-            accept_header_value="application/json",
-            http_headers=http_headers,
-            security=self.sdk_configuration.security,
-            timeout_ms=timeout_ms,
-        )
-
-        if retries == UNSET:
-            if self.sdk_configuration.retry_config is not UNSET:
-                retries = self.sdk_configuration.retry_config
-
-        retry_config = None
-        if isinstance(retries, utils.RetryConfig):
-            retry_config = (retries, ["429", "500", "502", "503", "504"])
-
-        http_res = await self.do_request_async(
-            hook_ctx=HookContext(
-                base_url=base_url or "",
-                operation_id="GetOneFileUploadUrl",
-                oauth2_scopes=[],
-                security_source=get_security_from_env(
-                    self.sdk_configuration.security, models.Security
-                ),
-            ),
-            request=req,
-            error_status_codes=["4XX", "500", "5XX"],
-            retry_config=retry_config,
-        )
-
-        response_data: Any = None
-        if utils.match_response(http_res, "200", "application/json"):
-            return utils.unmarshal_json(
-                http_res.text, Optional[models.GetOneFileUploadURLResponseBody]
-            )
-        if utils.match_response(http_res, "500", "application/json"):
-            response_data = utils.unmarshal_json(http_res.text, models.HonoAPIErrorData)
-            raise models.HonoAPIError(data=response_data)
-        if utils.match_response(http_res, "4XX", "*"):
             http_res_text = await utils.stream_to_text_async(http_res)
             raise models.APIError(
                 "API error occurred", http_res.status_code, http_res_text, http_res