+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -83,5 +83,5 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-29T18:00:56.880Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-31T06:21:55.956Z")
     r"""The date and time the resource was last updated"""