+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQHH0YGF1P8JG9PRB1Q5EMMD"
+        "01JQNDTED7Q668FGWAEE5YN3G1"
     )
     r"""The id of the resource"""