+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -86,7 +86,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQHH0YGGFZQSZ9XSQPVH3GCY"
+        "01JQNDTED8RDH1B5H21HPP8ZEQ"
     )
     r"""The id of the resource"""