+++ packages/orq-rc/RELEASES.md
@@ -959,3 +959,13 @@ Based on:
 - [python v3.3.0-rc.14] packages/orq-rc
 ### Releases
 - [PyPI v3.3.0-rc.14] https://pypi.org/project/orq-ai-sdk/3.3.0-rc.14 - packages/orq-rc
+
+## 2025-03-31 06:23:21
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.526.1 (2.563.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.3.0-rc.15] packages/orq-rc
+### Releases
+- [PyPI v3.3.0-rc.15] https://pypi.org/project/orq-ai-sdk/3.3.0-rc.15 - packages/orq-rc